Split test_mla.py into two files (#4216)
