+++ .github/workflows/pr-test-amd.yml
@@ -90,11 +90,11 @@ jobs:
       - name: MLA TEST
         timeout-minutes: 20
         run: |
-          docker exec -w /sglang-checkout/test/srt ci_sglang python3 test_mla.py TestMLA
+          docker exec -w /sglang-checkout/test/srt ci_sglang python3 test_mla.py
 
   finish:
     needs: [
-      accuracy-test-1-gpu-amd
+      accuracy-test-1-gpu-amd, mla-test-1-gpu-amd
     ]
     runs-on: ubuntu-latest
     steps:
